Feature: CodeReviewService 리팩터링,파일 WhiteList 필터링 기능 추가

+++ src/main/java/org/minturtle/careersupport/codereview/dto/CodeReviewRequest.java
@@ -13,5 +13,5 @@ import java.util.Map;
 public final class CodeReviewRequest {
     private String githubToken;
     private String repositoryName;
-    private Long prNumber;
+    private Integer prNumber;
 }
